M3: Add PMM frame refcounting, reclaim mmap regions on teardown
Replaces the manual `release_user_code_pfns` boolean policy (which guarded
against double-freeing user-code frames aliased over other mappings) with a
per-frame refcount array in the PMM. `alloc_frame` initializes a fresh
frame's count to 1; `inc_refcount(pfn)` bumps it when a frame gains an
additional owner (e.g. deliberate aliasing); `release_pfn` always decrements
and only actually frees the frame once the count reaches zero. Every
teardown call site now calls `release_pfn` unconditionally and lets the
refcount decide, instead of branching on a caller-supplied boolean.

Also extends `destroy_user_address_space` with a generic catch-all pass that
reclaims any present user mapping outside the fixed Code/Stack/Heap windows
(bounded to the user PML4 slot range so it can never touch the shared
low-memory identity map or higher-half kernel slots), so a future
per-process region (e.g. mmap) can't be silently leaked at teardown.

Updated the UEFI loader's PMM metadata reservation size to account for the
new refcount array (previously sized for the bitmap only), avoiding a
real-hardware regression on large-RAM systems.

Removed `release_user_code_pfns` from `TaskEntry`/`SpawnKind`/
`destroy_user_address_space_with_options` (now folded into
`destroy_user_address_space`/`_with_page_counts`) now that refcounting
makes the boolean redundant.

Added kernel/tests/pmm_test.rs coverage for single-owner and refcounted
(shared) frame release semantics and inc_refcount rejection of
free/out-of-range PFNs; added kernel/tests/vmm_test.rs coverage for the new
single-owner/shared-frame teardown semantics and the catch-all reclaim
pass; updated kernel/tests/pmm_uefi_test.rs's synthetic metadata-size
expectations for the new refcount array.

Also fixes a handful of pre-existing clippy lint violations in unrelated
test files (serial_deadlock_test, time_test, console_flush_lock_test,
syscall_dispatch_test, heap_test) that blocked a clean
`cargo clippy --tests -- -D warnings` run.

Verification: `cargo test` (kernel/, full suite: 350/350 cases across 37
test files passed), `cargo clippy` / `cargo clippy --tests -- -D warnings`
(kernel/), `cargo clippy -- -D warnings` (kaosldr_uefi/, kaosldr_64/), and
`cargo fmt --check` (workspace-wide) all clean.

Closes #15

+++ main64/kernel/tests/syscall_dispatch_test.rs
@@ -769,7 +769,7 @@ fn test_writable_user_buffer_rejects_read_only_page_after_boundary() {
         "a range crossing into a read-only page must be rejected"
     );
 
-    let ret = syscall::dispatch(SyscallId::GetTime as u64, (first_va + 128) as u64, 0, 0, 0);
+    let ret = syscall::dispatch(SyscallId::GetTime as u64, first_va + 128, 0, 0, 0);
     assert_eq!(
         ret,
         syscall::SYSCALL_OK,
